+++ src/app/howto/page.tsx
@@ -1,3 +1,9 @@
 export default function Page() {
   return <div>howto 페이지</div>;
 }
+
+// 컴포넌트를 Props로 받고
+// 대문제 Const로 받고 대문자 저걸 return 해.
+// render props
+// 근데 렌더 함수지만 본인이 멀 렌더하는 지를 몰라 => 상위 컴포넌트에서 진해아흔ㄴ거
+// Render props
